+++ src/services/demarchesSimplifiees/import.ts
@@ -1,5 +1,11 @@
 import { DSResponse } from "../../types/demarcheSimplifiee";
 import { DSPsychologist, Psychologist } from "../../types/psychologist";
+import {
+  requestDossiersWithAnnotations,
+  requestPsychologistsAcceptes,
+  requestPsychologistsById,
+  requestPsychologistsState,
+} from "./buildRequest";
 import { requestPsychologistsFor } from "./buildRequest";
 import parsePsychologists from "./parse-psychologists";
 
@@ -50,4 +56,50 @@ export const getPsychologistList = async (
   console.timeEnd(time);
 
   return results;
+};
+
+function fetchPsyById(ids: number[]) {
+  return Promise.all(
+    ids.map(async (id) => limit(() => requestPsychologistsById(id)))
+  );
+}
+
+function mapPsyData(psyData: any[]) {
+  return psyData.map((psy) => psy.dossier).flat();
+}
+
+export const getPsychologistFromListIds = async (
+  ids: number[]
+): Promise<Psychologist[]> => {
+  const psyData = await fetchPsyById(ids);
+  return await parsePsychologists(mapPsyData(psyData));
+};
+
+export const getPsychologistState = async (): Promise<
+  Partial<Psychologist>[]
+> => {
+  const time = `Fetch all psychologists state from DS (query id #${Math.random().toString()})`;
+
+  console.time(time);
+  const list = await getAllPsychologistList(requestPsychologistsState);
+  console.timeEnd(time);
+
+  return list.psychologists.map((dossier) => ({
+    archived: dossier.archived,
+    id: dossier.number,
+    state: dossier.state,
+  }));
+};
+export const getDossiersInConstruction = async (): Promise<
+  DSPsychologist[]
+> => {
+  const time = `Fetching all psychologists folders in construction from DS`;
+
+  console.time(time);
+  const list = await getAllPsychologistList((cursor) =>
+    requestDossiersWithAnnotations(cursor)
+  );
+  console.timeEnd(time);
+
+  return list.psychologists;
 };